requirements.txt
datasets
diffusers
accelerate
peft
PyYAML
Pillow
tqdm
pydantic
torch_fidelity
typer[all]
# torch, torchvision, torchaudio will be installed separately based on CUDA version
# xformers is optional

# Dependencies for Benchmark Evaluation (calc_clip.py, etc.)
torchmetrics
opencv-python-headless
clip-score